fix: don't reference testing package from main module (#11)
The `values_generator.go` code was importing "testing", which means code
that is using this module needs to pull in "testing". I believe this was
an oversight, as it's not a best practice.

This is technically a breaking API change, since it's removing API
methods from the public interface of the module. I'd like to label it as
a fix rather than rev'ing to `v2` because:
1. I can't find any references to it on Github 
2. It wasn't intended to be part of the public interface